+++ lib/cluster.rb
@@ -58,6 +58,7 @@ require './lib/cluster/config_checks/storage'
 require './lib/cluster/config_checks/json_format'
 require './lib/cluster/config_checks/bucket_configs'
 require './lib/cluster/config_checks/real_users'
+require './lib/cluster/config_checks/secrets_ips'
 
 module Cluster
   class JSONFormatError < StandardError; end